requirements.txt.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
umap-learn
lime
